// src/app/model/suggestions.ts
export class Suggestions {
  rent: number;
  savings: number;
  phoneBills: number;
  sodexo: number;
  otherInsurance: number;
  insurance: number;
  NPSPension: number;
  pf: number;
  basic: number;
  hra: number;
  stand: number;
  phone: number;
  spl: number;
  gratuity: number;
  pt: number;
  esi: number;
  gross: number;
}